Avoid hard coding chunkMaxLogicalBytes in parquet writer

+++ lib/trino-parquet/src/main/java/io/trino/parquet/writer/ParquetWriterOptions.java
@@ -40,7 +40,7 @@ public class ParquetWriterOptions
         this.batchSize = batchSize;
     }
 
-    public long getMaxRowGroupSize()
+    public int getMaxRowGroupSize()
     {
         return maxRowGroupSize;
     }
